Fix the bug
I forgot that y axis is upward direction in unity.
So I swap the y value and z value for chrom[i] to get the right vector.

Also, I changed theta and phi to get a proper randomized angle.
if theta is over Math.PI/2, the bullet is goind downward which means it does not reach the target.
So the range of theta is from PI/6 to PI/2 (you can see the bullet is going upward in that range)